\name{idgaf-package}
\alias{idgaf-package}
\docType{package}
\title{.}
\description{
  This package contains all public GitHub commits containing the word
  "fuck" in the commit message, from 3/11/2012 to 7/24/2014.
}
\details{
\tabular{ll}{
  Package: \tab idgaf\cr
  Type: \tab Package\cr
  License: \tab BSD 2-Clause\cr
}
}
\references{
  \url{https://docs.google.com/spreadsheets/d/1NDKNmTS25Ijqay3BjB6c6N1MeunFgDWQcqijCOSRLVk/edit?usp=sharing}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
