\name{fuck-data}
\alias{fuck-data}
\alias{fuck}

\docType{data}

\title{The Full GitHub "fuck" Dataset}

\description{
  All public GitHub commits containing the word
  "fuck" in the commit message, from 3/11/2012 to 7/24/2014.
}
\details{
  The data is stored in a dataframe with the following columns:
\tabular{ll}{
  commit  \tab  The content of the commit message. \cr
  url     \tab  The url of the commit. \cr
  user    \tab  The user who made the commit. \cr
  repo    \tab  The repository that was committed to. \cr
  date    \tab  The day the commit was made. \cr
  time    \tab  The time the commit was made 
}
}

