\name{fucklang-data}
\alias{fucklang-data}
\alias{fucklang}

\docType{data}

\title{The Number of "fucks" (and "shits") by Programming Language}

\description{
  A breakdown of some popular programming languages and the prevalence
  of their pushes, "fucks", and "shits".
}
\details{
  The data is stored in a dataframe with the following columns:
\tabular{ll}{
  language        \tab  Programming language. \cr
  pct_total_swear \tab  Percentage of swears among all pushes. \cr
  pushes          \tab  Total number of pushes. \cr
  fucks           \tab  Total number of pushes containing "fuck". \cr
  shits           \tab  Total number of pushes containing "shit".
}
}

